requirements.txt
Faker>=37.0.2
pytest-base-url>=2.1.0
pytest-playwright>=0.7.0
PyYAML>=6.0.2
pytest-html>=4.1.1